Add Activities and DataEventFetcher
Adding a MainView Activity which is a temporary main window for the application.
Adding a ConfigurationView Activity to configure the systems: it currently
configures the background updated mechanism implemented by the EventReader.
Adding a DataEventFetcher mechanism (with abstract class and factory) to
implement different web access policy to retrieve event data.
Adding a mock DataEventFetcherHTTP.