Implement DataEventFetcherHTTP
Implementing a more realistic DataEventFetcherHTTP to retrieve the event data (in json format) from the faitango.it website.
Adding a (test) button on the main activity to manually force synchronization (fetching data).
Adding a test button to manually clean the DB.
Adding an EventFilter to restrict the synchronization, adding the generation of the URI query as well.