Intent passing search params to EventReader
Using the Intent extras mechanism to pass the EventFilter object
to the EventReader class.
EventFilter is now a Parcelable, as well as EventType.
EventFilter is partially created from preferences (search parameters).
FIX province values (must be of type PI for Pisa) and add the
"any province" and "any region" options in the preferences.