Fix EventFilter through Intent
The EventFilter is now properly passed from the EventReaderAlarm
using the Parcelable mechanism.
Fixing the problem related to the PendingIntent mechanism required
to have periodic and non-periodic activations in EventReaderAlarm.
The issues was that the PendingIntent.getService(...) had to be
called with the proper FLAG in order to refresh the Intent and its
extras part (thus the EventFilter).
Other minor changes and fixes.