EventReader change and Startup sync
EventReader is no longer a singleton class.
It now exports the isExetuting§() method as a synchronized one.
TODO: Check if the mutex condition is still consistent.
Add start-up synchronization which might be performed when the
application is launched, according to the preferences.
Add (possible) waiting message to the EventReaderAsyncTask
constructor.